+++ src/screens/HomeScreen.js
@@ -1,4 +1,4 @@
-import { StyleSheet, View } from "react-native";
+import { StyleSheet, ScrollView } from "react-native";
 import VolunteerOpportunity from "../components/VolunteerOpportunity";
 import OtherOpportunities from "../components/OtherOpportunities";
 import Heading from "../components/Heading";
@@ -8,7 +8,7 @@ export default function HomeScreen({ navigation }) {
   console.log("Switched");
 
   return (
-    <View style={styles.container}>
+    <ScrollView style={styles.container}>
       <Heading>Volunteer Opportunities</Heading>
       <VolunteerOpportunity navigation={navigation} />
       <VolunteerOpportunity navigation={navigation} />
@@ -17,7 +17,7 @@ export default function HomeScreen({ navigation }) {
       <OtherOpportunities />
       <Heading>Websites</Heading>
       <Websites />
-    </View>
+    </ScrollView>
   );
 }
